Refactor: create new custom hook to parse logs
