+++ app/(dashboard)/dashboard/log-view.tsx
@@ -1,58 +1,15 @@
 'use client'
 
-import { useRef, useMemo, useState } from 'react'
+import { useRef, useState } from 'react'
 import { Loader2, ChevronRight, ChevronDown } from 'lucide-react'
-import { LogGroup, LogViewProps } from './types'
-
-
+import { LogViewProps } from './types'
+import { useLogGroups } from '@/hooks/use-log-groups'
 
 export function LogView({ logs, error, isLoading }: LogViewProps) {
   const logContainerRef = useRef<HTMLDivElement>(null)
   const [expandedGroups, setExpandedGroups] = useState<Record<string, boolean>>({})
 
-  const parsedLogs = useMemo(() => {
-    if (!logs) return [];
-
-    const groups: LogGroup[] = [];
-    let currentGroup: LogGroup | null = null;
-    const lines = logs.split('\n');
-
-    for (let i = 0; i < lines.length; i++) {
-      const line = lines[i].replace(/^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}\.\d+Z\s*/, '');
-      if (line.startsWith('File:')) {
-        if (currentGroup) {
-          groups.push(currentGroup);
-        }
-        currentGroup = {
-          id: `group-${groups.length}`,
-          name: line.trim(),
-          logs: []
-        };
-      } else if (currentGroup) {
-        currentGroup.logs.push(line);
-      } else {
-        if (!groups.length || groups[groups.length - 1].name !== 'Other') {
-          groups.push({ id: `group-${groups.length}`, name: 'Other', logs: [] });
-        }
-        groups[groups.length - 1].logs.push(line);
-      }
-    }
-
-    if (currentGroup) {
-      groups.push(currentGroup);
-    }
-
-    // Add line numbers and trim group names
-    return groups.map(group => ({
-      ...group,
-      name: group.name
-        .replace(/^File:\s*/, '')
-        .replace(/^.*?_/, '')
-        .replace(/\.txt$/, '')
-        .split('/')[0],
-      logs: group.logs.map((log, index) => `${(index + 1).toString().padStart(4, ' ')} | ${log}`)
-    }));
-  }, [logs]);
+  const parsedLogs = useLogGroups(logs);
 
   const toggleGroup = (groupId: string) => {
     setExpandedGroups(prev => ({
